requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coverage==5.4
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-crontab==0.7.1
django-heroku==0.3.1
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
drf-social-oauth2==1.0.8
googlemaps==4.4.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
